serial -> smallserial

+++ src/Postgres.Marula.DatabaseAccess/SqlScripts/ScriptFiles/06-parameter-values-history-table.sql
@@ -5,7 +5,7 @@ set search_path to SYSTEM_SCHEMA_NAME_TO_REPLACE;
 create table VALUES_HISTORY_TABLE_NAME_TO_REPLACE
 (
 	id                    bigserial                   primary key,
-	parameter_id          int                         not null references calculated_parameters(id),
+	parameter_id          smallint                    not null references calculated_parameters(id),
 	calculated_value      non_empty_string            not null,
 	status                STATUS_ENUM_NAME_TO_REPLACE not null,
 	calculation_date_time timestamp                   not null default now()
